fix: move accoutn tyep in separate type file
