+++ apps/bilan-carbone/src/app/(dashboard)/etudes/[id]/(organization)/cadrage/modifier-contributeur/[accountId]/page.tsx
@@ -3,7 +3,8 @@ import { StudyProps } from '@/components/hoc/withStudy'
 import withStudyDetails from '@/components/hoc/withStudyDetails'
 import EditStudyContributorPage from '@/components/pages/EditStudyContributor'
 import NotFound from '@/components/pages/NotFound'
-import { AccountWithUser, getAccountById } from '@/db/account'
+import { getAccountById } from '@/db/account'
+import { AccountWithUser } from '@/types/account.types'
 import { getAccountRoleOnStudy, hasEditionRights } from '@/utils/study'
 import { redirect } from 'next/navigation'
