+++ apps/bilan-carbone/src/components/study/EmissionSourceEditorChip.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import StyledChip from '@/components/base/StyledChip'
-import { AccountWithUser } from '@/db/account'
 import type { FullStudy } from '@/db/study'
+import { AccountWithUser } from '@/types/account.types'
 import { getDisplayedRoleOnStudy } from '@/utils/study'
 import { accountWithUserToUserSession } from '@/utils/userAccounts'
 import SpaIcon from '@mui/icons-material/Spa'
