+++ apps/bilan-carbone/src/components/study/rights/NewStudyContributorForm.tsx
@@ -4,7 +4,6 @@ import Form from '@/components/base/Form'
 import LoadingButton from '@/components/base/LoadingButton'
 import MultiplePosts from '@/components/emissionFactor/Form/MultiplePosts'
 import { FormTextField } from '@/components/form/TextField'
-import { AccountWithUser } from '@/db/account'
 import { getOrganizationVersionAccounts } from '@/db/organization'
 import type { FullStudy } from '@/db/study'
 import { useServerFunction } from '@/hooks/useServerFunction'
@@ -19,6 +18,7 @@ import {
   NewStudyContributorCommandValidation,
 } from '@/services/serverFunctions/study.command'
 import { useAppEnvironmentStore } from '@/store/AppEnvironment'
+import { AccountWithUser } from '@/types/account.types'
 import { SubPost } from '@abc-transitionbascarbone/db-common/enums'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useTranslations } from 'next-intl'
