+++ apps/bilan-carbone/src/services/permissions/emissionFactor.ts
@@ -1,4 +1,4 @@
-import type { AccountWithUser } from '@/db/account'
+import type { AccountWithUser } from '@/types/account.types'
 import type { EmissionFactor } from '@abc-transitionbascarbone/db-common'
 import { Import } from '@abc-transitionbascarbone/db-common/enums'
 import { isFromEmissionFactorOrganization } from '../serverFunctions/emissionFactor'
