+++ apps/bilan-carbone/src/types/account.types.ts
@@ -1,0 +1,6 @@
+import { type Account, type User, Environment } from '@repo/db-common'
+
+export type AccountWithUser = Account & {
+  user: User
+  organizationVersion: { organizationId: string; environment: Environment }
+}
